requirements-mlops.txt
dvc>=3.50.0
mlflow>=2.14.0
evidently>=0.4.39,<0.5.0
pandas>=2.2.0
pyyaml>=6.0.0
apache-airflow>=2.8.0
apache-airflow-providers-python>=4.10.0
